chnaged the media sequence from long to int

+++ src/main/java/io/lindstrom/m3u8/model/MediaPlaylist.java
@@ -12,7 +12,7 @@ public interface MediaPlaylist extends Playlist {
     int targetDuration();
 
     @Value.Default
-    default long mediaSequence() {
+    default int mediaSequence() {
         return 0;
     }
